servic-app/requirements.txt
asgiref
attrs
Django
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
inflection
jsonschema
jsonschema-specifications
pillow
pip
psycopg2
PyJWT
python-dotenv
PyYAML
referencing
rpds-py
setuptools
sqlparse
typing_extensions
tzdata
uritemplate